Сделал новый модуль utils и перенёс туда функции
